move es bulk endpoint handling
